// frontend/src/graphql/models/Vehicle.ts
export type Vehicle = {
  id: number;
  image: File;
  model: string;
  brand: string;
  color: string;
  price: number;
  year: number;
  rating: number;
  category: string;
};
export type VehicleResponse = {
  id: number;
  image: string;
  model: string;
  brand: string;
  color: string;
  price: number;
  year: number;
  rating: number;
  category: string;
};